[FIX] #5 For better screens compatibility...
Continuation of previous commit, use BGA placeOnObjectPos to get better positionning.
Previous commit was OK on large screen and really chaotic on small screens (real or virtual mobiles, even rotated).
PS : it is not perfect placement (as 2 previous commit ago), but at least it works on mobile and PC where are loads to select are in the frame.
PS : https://boardgamearena.com/bug?id=104115